feat: Add ReportPage, SignUpApprovalPage, TradersPage, and UsersPage components
- Implemented ReportPage for displaying monthly performance reports with stats and a table.
- Created SignUpApprovalPage for managing user registration requests with approval and rejection functionality.
- Developed TradersPage to showcase traders with their details and images.
- Added UsersPage for displaying users directory with filtering options based on roles and status management.
