+++ client/src/user/superadmin/EmployeesPage/AddEmployeeModal.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import Button from "../../components/ui/Button";
+import Button from "../../../components/ui/Button";
 
 const AddEmployeeModal = ({ onClose, onAdd }) => {
   const [form, setForm] = useState({
