+++ client/src/user/superadmin/EmployeesPage/ConfirmRemoveModal.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Button from "../../components/ui/Button";
+import Button from "../../../components/ui/Button";
 
 const ConfirmRemoveModal = ({ onClose, onConfirm }) => (
   <div className="fixed inset-0 flex items-center justify-center bg-black/50 backdrop-blur-sm z-50">
