+++ client/src/user/superadmin/EmployeesPage/EmployeeCard.jsx
@@ -1,7 +1,7 @@
 import React from "react";
-import Button from '../../components/ui/Button';
-import Badge from "../../components/ui/Badge";
-import { Card } from "../../components/ui/Card";
+import Button from '../../../components/ui/Button';
+import Badge from "../../../components/ui/Badge";
+import { Card } from "../../../components/ui/Card";
 import { EMPLOYEE_STATUS_CONFIG } from "./employeeConstants"; // Import from new file
 
 const EmployeeCard = ({ employee, onAssign, onRemoveClick, isLoading }) => {
