+++ client/src/user/superadmin/EmployeesPage/EmployeeHeader.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Button from "../../components/ui/Button";
+import Button from "../../../components/ui/Button";
 
 const EmployeeHeader = ({ onAddEmployee }) => {
   return (
