+++ client/src/user/superadmin/Manufacturer/Payment.jsx
